show existing pages in the new layout
