+++ src/containers/Layout/MainLayout.js
@@ -6,6 +6,7 @@ import ReduxToastr from 'react-redux-toastr'
 import 'react-redux-toastr/lib/css/react-redux-toastr.min.css'
 // styles
 import CssBaseline from '@material-ui/core/CssBaseline'
+import ChluLayout from './ChluLayout'
 
 class MainLayout extends Component {
   static propTypes = {
@@ -17,7 +18,7 @@ class MainLayout extends Component {
 
     return (
       <CssBaseline>
-        {children}
+        <ChluLayout>{children}</ChluLayout>
         <ReduxToastr
           timeOut={5000}
           preventDuplicates
